+++ cypress/fixtures/completely-broken-routes.js
@@ -1,1 +1,0 @@
-lkewjflkjadsf // eslint-disable-line