+++ cypress/fixtures/plugins/plugin-bar/filters.js
@@ -1,6 +1,0 @@
-// local dependencies
-const { addFilter } = require('@nowprototypeit/govuk').views
-
-addFilter('bar__link',
-  (content, url) => `<a href="${url || '#'}">${content}</a>`,
-  { renderAsHtml: true })