+++ cypress/fixtures/plugins/plugin-foo/filters.js
@@ -1,2 +1,0 @@
-const { addFilter } = require('@nowprototypeit/govuk').views
-addFilter('foo__strong', (content) => `<strong>${content}</strong>`, { renderAsHtml: true })