+++ cypress/fixtures/plugins/plugin-foo/functions.js
@@ -1,2 +1,0 @@
-const { addFunction } = require('@nowprototypeit/govuk').views
-addFunction('fooEmphasize', (content) => `<em>${content}</em>`, { renderAsHtml: true })