+++ cypress/fixtures/plugins/plugin-foo/scripts/foo-submodule.js
@@ -1,3 +1,0 @@
-export default function fooSubmodule (x, y) {
-  return x + y
-}