+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
  - Striped out GOV.UK References
  - Put in Now Prototype It branding (unstyled)
+ - New Plugin Details page
 
 To see historic changes before we forked the GOV.UK Prototype Kit check out their changelog from that time: https://github.com/alphagov/govuk-prototype-kit/blob/2fedd466c8628519be54aca3ac86695dfd241ae1/CHANGELOG.md